fix: correct id type to integer as parameter in method fetchArtworkDetail of HarvardApiClient

+++ src/main/java/com/northcoders/exhibition_curation_platform/service/HarvardApiClient.java
@@ -54,7 +54,7 @@ public class HarvardApiClient {
 
     }
 
-    public Artwork fetchArtworkDetail (Long id, Artwork existingArtwork){
+    public Artwork fetchArtworkDetail (int id, Artwork existingArtwork){
         String url = BASE_URL +
                 "/" + id +
                 "?apikey=" + ApiKeyManager.getHAMApiKey();
